split lektor.environment to reduce cycles
